feat: refactor query routing to enhance provider selection and scoring
- Implemented a new scoring mechanism in QueryRouter using performance metrics and confidence scores.
- Removed the EnhancedQueryRouter and integrated its functionality directly into QueryRouter.
- Added budget-aware selection and improved decision explanation generation.
- Introduced comprehensive tests for the new routing logic and scoring system.

This update simplifies the routing logic while enhancing the scoring capabilities, adhering to the KISS principle by consolidating functionality into a single class
